1318 - move partners folder to admin
