+++ src/components/admin/partners/PartnersContent.tsx
@@ -2,7 +2,7 @@ import { Box, Container, Grid } from '@mui/material'
 import Image from 'next/image'
 import { Fragment } from 'react'
 import { useTranslation } from 'react-i18next'
-import useMobile from '../../common/hooks/useMobile'
+import useMobile from '../../../common/hooks/useMobile'
 import { organizationLogos, partners, techAndMarketingPartners } from './helpers/partnersData'
 import { SectionTitle, StyledGrid } from './Partners.styled'
 import PartnersSlider from './PartnersSlider'
