+++ src/components/admin/partners/PartnersSlider.tsx
