+++ src/components/admin/partners/helpers/giversData.ts
