+++ src/components/admin/partners/helpers/guarantsData.ts
