+++ src/components/admin/partners/helpers/mediaAboutUsData.ts
