+++ src/components/admin/partners/helpers/partnersData.ts
