+++ src/components/admin/partners/helpers/tabs.ts
